dynamically update number of beats with selector, TODO match indexes and use better name WIP
